Move board and util logic to specific modules.
